# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMS CRM API Documentation",
        "description": "API documentation for the HRMS CRM application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://hrmscrm.test",
            "description": "Local Development Server"
        },
        {
            "url": "https://ymhrms.girafdev.com",
            "description": "Demo Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/socials": {
            "get": {
                "tags": [
                    "Social Media"
                ],
                "summary": "Get social media feed",
                "description": "Get social feed (published posts).",
                "operationId": "3850a86d292ae921d7cccec027f76274",
                "responses": {
                    "200": {
                        "description": "Social feed retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Social Media"
                ],
                "summary": "Create a new social post",
                "description": "Create a new social post.",
                "operationId": "390cacd8148dfb3a12b129ac0c01884a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Social post created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}